refactor: split idiomatic_patterns and exhaustive_intrinsics across crates
idiomatic_patterns: Patterns 1-6, 8-9 (archmage macro patterns using
raw intrinsics) stay in archmage. Pattern 7 (magetypes type operations)
stays in magetypes.

exhaustive_intrinsics: Token availability, summon behavior, naming, and
dispatch pattern tests stay in archmage. implementation_name() tests
for all magetypes SIMD types stay in magetypes.
